requirements.txt
fastapi
uvicorn
python-dotenv
openai
pymysql
cryptography
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
